requirements.txt
django~=4.0
djangorestframework==3.13.1
psycopg2-binary
django-filter~=21.1